// app-angular/src/api/se/model/storageEntryDto.ts
/**
 * Storage Explorer Embedded RESTful API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Reference } from './reference';
import { StorageEntryType } from './storageEntryType';


export interface StorageEntryDto { 
    type: StorageEntryType;
    uri: string;
    schema: string;
    name: string;
    /**
     * Type identifier returned for OBJECTs. 
     */
    typeName?: string;
    /**
     * Current sequence value returned for SEQUENCEs. 
     */
    seqVal?: number;
    /**
     * Host entry URI returned for scoped entries. 
     */
    scopeHost?: string;
    references: Array<Reference>;
}